refactor(mcp-server): split proxy enforcement pdp
