+++ crates/assay-mcp-server/src/proxy/enforce/tests.rs
@@ -1,0 +1,12 @@
+use super::allowance::target_digest;
+use super::credential_scope::{credential_scope_gate, scope_covers, ScopeCoverage};
+use super::*;
+use anyhow::Result;
+use serde_json::{json, Value};
+use std::io::Write;
+
+mod drift;
+mod fixtures;
+mod pdp;
+mod policy;
+mod records;
